fix(meetup): 모임 만들기 메타데이터와 클라이언트 컴포넌트 충돌 해결
